+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add label to variables
+- Number variable - support min/max
 
 ## [7.1.0] - 2023-05-18